Comment 페이지 초기세팅 완료
