+++ src/Pages/Comment/Components/ReplyBox/ReplyBox.scss
@@ -1,0 +1,48 @@
+.rereplyContainer {
+  margin-top: 20px;
+  margin-left: 30px;
+
+  .name {
+    margin-left: 7px;
+    margin-bottom: 10px;
+    font-size: 12px;
+    font-weight: bold;
+  }
+
+  .textBox {
+    display: inline-block;
+    background-color: rgb(240, 241, 244);
+    font-size: 12px;
+    margin-bottom: 10px;
+    padding: 14px;
+    border-radius: 0px 6.4vw 6.4vw;
+  }
+
+  .commentInfo {
+    margin-left: 15px;
+    margin-bottom: 20px;
+
+    .time {
+      font-size: 13px;
+      color: lightslategray;
+      margin-right: 10px;
+    }
+
+    .heartBtn {
+      background-color: white;
+      color: gray;
+      font-size: 13px;
+      font-weight: bold;
+
+      .heart {
+        display: inline-block;
+        font-size: 17px;
+        margin-bottom: -4px;
+      }
+    }
+
+    .likeCount {
+      margin-left: 5px;
+    }
+  }
+}
