+++ src/Pages/Main/Main.scss
@@ -1,4 +1,4 @@
-@import "../../Styles/common.scss";
+@import '../../Styles/common.scss';
 
 .mainContainer {
   background-color: $theme_color;
